Правки 6-8

+++ src/index.js
@@ -19,7 +19,13 @@ class CommentApp extends React.Component {
 	        comments: comments || defaultComments,
 	        newAuthor: "",
 	        newComment: ""
-    	};
+    	};  	
+
+    	this.state.comments.map((comment, i) => {
+    		this.remove = this.remove.bind(this, i);
+    	})	
+
+		
 
 		//localStorage.setItem('comments', '');
 
@@ -46,24 +52,36 @@ class CommentApp extends React.Component {
 		const fullDate = "" + dayToday+"-"+ monthToday + "-" + yearToday;
 		//console.log('now');
 
-		comments.push({
-			author: this.state.newAuthor,
-			date: fullDate,
-			text: this.state.newComment
-		});
+		if ((this.state.newAuthor.replace(/\s+/g," ").trim() != '') && (this.state.newComment.replace(/\s+/g," ").trim() != '')) {
 
-		localStorage.setItem('comments', JSON.stringify(this.state.comments));
+			comments.push({
+				author: this.state.newAuthor,
+				date: fullDate,
+				text: this.state.newComment
+			});
 
-		this.setState({ 
-			comments, 
-			newAuthor: '',
-			newComment: ''
-		});
+			localStorage.setItem('comments', JSON.stringify(this.state.comments));
+
+			this.setState({ 
+				comments, 
+				newAuthor: '',
+				newComment: ''
+			});
+		}	
 		
+	}
+
+	onChangeName(ev) {
+		this.setState({ newAuthor: ev.target.value});
+	}
+
+	onChangeComment(ev) {
+		this.setState({ newComment: ev.target.value});
 	}
 
 	render() {
 		// JSX-синтаксис
+		const { comments } = this.state;
 
 		return (
 
@@ -72,14 +90,15 @@ class CommentApp extends React.Component {
 
 				<ul>
 					{
-						this.state.comments.map((comment, i) => {
+						comments.map((comment, i) => {
+							const { author, text, date } = comment;
 							return (
 								<CommentItem
 									key = {i}
-									author = {comment.author}
-									text = {comment.text}
-									date = {comment.date}							
-									remove = {this.remove.bind(this, i)}
+									author = {author}
+									text = {text}
+									date = {date}							
+									remove = { this.remove }
 							/>
 							)
 						})
@@ -94,9 +113,7 @@ class CommentApp extends React.Component {
 						required
 						placeholder = "Введите имя"
 						value = {this.state.newAuthor}
-						onChange = {ev => {
-							this.setState({ newAuthor: ev.target.value});
-						}}
+						onChange = {ev => this.onChangeName(ev)}
 						
 					/>
 
@@ -105,20 +122,14 @@ class CommentApp extends React.Component {
 						rows="10" 
 						cols="45" 
 						name="comment"
-						onChange = {ev => {
-							this.setState({ newComment: ev.target.value});
-						}}
+						onChange = {ev => this.onChangeComment(ev)}
 						
 					>
 						{this.state.newComment}			  
 					</textarea>					
 
-					<button
-						onClick ={ev => {	
-							if ((this.state.newAuthor.replace(/\s+/g," ").trim() != '') && (this.state.newComment.replace(/\s+/g," ").trim() != '')) {
-								this.addNewComment();
-							}				    							
-						}}
+					<button						
+						onClick={() => this.addNewComment()}
 					>
 					Добавить
 					</button>
